Move ZToaster into its own submodule so it can be built independently
