+++ .github/workflows/release.yml
@@ -50,9 +50,9 @@ jobs:
           TAG_NAME=${GIT_REF_NAME##*/}
           REVISION_NAME=${TAG_NAME:1}
 
-          mvn -B deploy "-Drevision=$REVISION_NAME"
+          mvn -B deploy -pl quartzlib -am "-Drevision=$REVISION_NAME"
 
-          JAR_PATH=$(ls ./target/*.jar | tail -n 1)
+          JAR_PATH=$(ls ./quartzlib/target/*.jar | tail -n 1)
           JAR_NAME=$(basename $JAR_PATH)
           echo "::set-output name=jar_path::$JAR_PATH"
           echo "::set-output name=jar_name::$JAR_NAME"
