+++ ztoaster/src/main/java/fr/zcraft/ztoaster/Toast.java
