+++ ztoaster/src/main/java/fr/zcraft/ztoaster/ToasterSidebar.java
