+++ ztoaster/src/main/java/fr/zcraft/ztoaster/commands/AddCommand.java
