+++ ztoaster/src/main/java/fr/zcraft/ztoaster/commands/OpenCommand.java
