+++ ztoaster/src/test/java/fr/zcraft/ztoaster/core/QuartzLibTest.java
@@ -1,8 +1,9 @@
-package fr.zcraft.quartzlib.core;
+package fr.zcraft.ztoaster.core;
 
 import static org.junit.Assert.assertSame;
 
-import fr.zcraft.quartzlib.MockedToasterTest;
+import fr.zcraft.quartzlib.core.QuartzLib;
+import fr.zcraft.ztoaster.MockedToasterTest;
 import org.junit.Test;
